Refactor around test asset in tests of Zend\Log
