+++ tests/Zend/Log/TestAsset/CustomSyslogWriter.php
@@ -1,0 +1,12 @@
+<?php
+namespace ZendTest\Log\TestAsset;
+
+use Zend\Log\Writer\Syslog as SyslogWriter;
+
+class CustomSyslogWriter extends SyslogWriter
+{
+    public function getFacility()
+    {
+        return $this->_facility;
+    }
+}
