+++ tests/Zend/Log/TestAsset/MockDbAdapter.php
@@ -1,0 +1,12 @@
+<?php
+namespace ZendTest\Log\TestAsset;
+
+class MockDbAdapter
+{
+    public $calls = array();
+
+    public function __call($method, $params)
+    {
+        $this->calls[$method][] = $params;
+    }
+}
