+++ tests/Zend/Log/TestAsset/MockHttpResponse.php
@@ -1,0 +1,52 @@
+<?php
+namespace ZendTest\Log\TestAsset;
+
+use Zend\Controller\Response\Http as HttpResponse;
+
+class MockHttpResponse extends HttpResponse
+{
+    public function canSendHeaders($throw = false)
+    {
+        return true;
+    }
+
+    public function verifyHeaders($headers)
+    {
+        $response_headers = $this->getHeaders();
+        if (!$response_headers) {
+            return false;
+        }
+
+        $keys1 = array_keys($headers);
+        sort($keys1);
+        $keys1 = serialize($keys1);
+
+        $keys2 = array();
+        foreach ($response_headers as $header ) {
+            $keys2[] = $header['name'];
+        }
+        sort($keys2);
+        $keys2 = serialize($keys2);
+
+        if ($keys1 != $keys2) {
+            return false;
+        }
+
+        $values1 = array_values($headers);
+        sort($values1);
+        $values1 = serialize($values1);
+
+        $values2 = array();
+        foreach ($response_headers as $header ) {
+            $values2[] = $header['value'];
+        }
+        sort($values2);
+        $values2 = serialize($values2);
+
+        if ($values1 != $values2) {
+            return false;
+        }
+
+        return true;
+    }
+}
