+++ tests/Zend/Log/TestAsset/StringObject.php
@@ -1,0 +1,10 @@
+<?php
+namespace ZendTest\Log\TestAsset;
+
+class StringObject
+{
+    public function __toString()
+    {
+        return 'Hello World';
+    }
+}
